backend/pyproject.toml
[project]
name = "mel-store-backend"
version = "0.1.0"
description = "FastAPI + LangGraph backend for Htech Store"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "psycopg[binary]>=3.2.0",
  "pgvector>=0.3.5",
  "langchain>=0.3.9",
  "langchain-postgres>=0.0.12",
  "langchain-google-genai>=2.0.5",
  "langchain-openai>=0.2.8",
  "langgraph>=0.2.56",
  "python-dotenv>=1.0.1",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
